# internal/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Tunploy API",
    "version": "1.2",
    "description": "Manage devices and servers on your Tunploy VPN from your own programs, and hear about what happens to them.\n\nSend an API key from Settings → API keys as `Authorization: Bearer tp_…`. Each key only does what its scopes allow. Requests are limited to 10 a second per key, with bursts up to 60. Keep keys on your server, never in an app.\n\nErrors always look like `{\"error\": {\"code\": \"…\", \"message\": \"…\", \"fields\": {…}}}`; `fields` names the inputs a validation error is about."
  },
  "servers": [
    {
      "url": "/api/v1"
    }
  ],
...
  "paths": {
...
    "/groups/{external_id}": {
      "get": {
        "operationId": "getGroup",
        "summary": "Get a group",
        "tags": [
          "Groups"
        ],
        "security": [
          {
            "apiKey": []
          }
        ],
        "x-scope": "devices:read",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Group"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          }
        },
        "description": "Every device with this `external_id`, oldest first, and their usage this month together.\n\nNeeds the `devices:read` scope.",
        "parameters": [
          {
            "name": "external_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "maxLength": 255
            },
            "description": "The external_id the devices share, URL-encoded."
          }
        ]
      },
      "patch": {
        "operationId": "updateGroup",
        "summary": "Change every device in a group",
        "tags": [
          "Groups"
        ],
        "security": [
          {
            "apiKey": []
          }
        ],
        "x-scope": "devices:write",
        "responses": {
          "200": {
            "description": "The group after the change.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Group"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "422": {
            "$ref": "#/components/responses/Invalid"
          },
          "502": {
            "description": "`apply_failed`: saved, but a running tunnel could not be updated.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          }
        },
        "description": "Sets the fields sent on every device in the group: renew a plan with `expires_at`, or end it with `enabled: false`. Every device is checked first, so a value that is wrong for one changes none.\n\nNeeds the `devices:write` scope.",
        "parameters": [
          {
            "name": "external_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "maxLength": 255
            },
            "description": "The external_id the devices share, URL-encoded."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "metadata": {
                    "type": [
                      "object",
                      "null"
                    ],
                    "additionalProperties": true,
                    "description": "Any JSON object up to 4 KB; null clears it."
                  },
                  "enabled": {
                    "type": "boolean"
                  },
                  "data_limit": {
                    "type": "integer",
                    "minimum": 0
                  },
                  "limit_period": {
                    "$ref": "#/components/schemas/LimitPeriod"
                  },
                  "expires_at": {
                    "type": [
                      "string",
                      "null"
                    ],
                    "format": "date-time",
                    "description": "null clears it."
                  },
                  "speed_limit": {
                    "type": "integer",
                    "minimum": 0,
                    "maximum": 10000000,
                    "description": "kbit/s (1000 bits a second), download and upload each; 0 means no limit."
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "operationId": "deleteGroup",
        "summary": "Delete every device in a group",
        "tags": [
          "Groups"
        ],
        "security": [
          {
            "apiKey": []
          }
        ],
        "x-scope": "devices:write",
        "responses": {
          "204": {
            "description": "Deleted."
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "502": {
            "description": "`apply_failed`: deleted, but a running tunnel could not be updated.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          }
        },
        "description": "Needs the `devices:write` scope.",
        "parameters": [
          {
            "name": "external_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "maxLength": 255
            },
            "description": "The external_id the devices share, URL-encoded."
          }
        ]
      }
    },
    "/groups/{external_id}/usage/reset": {
      "post": {
        "operationId": "resetGroupUsage",
        "summary": "Reset usage for a group",
        "tags": [
          "Groups"
        ],
        "security": [
          {
            "apiKey": []
          }
        ],
        "x-scope": "devices:write",
        "responses": {
          "200": {
            "description": "The group after the reset.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Group"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "502": {
            "description": "`apply_failed`: reset, but a running tunnel could not be updated.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          }
        },
        "description": "Resets every device in the group, as `POST /devices/{id}/usage/reset` does for one.\n\nNeeds the `devices:write` scope.",
        "parameters": [
          {
            "name": "external_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "maxLength": 255
            },
            "description": "The external_id the devices share, URL-encoded."
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "schema": {
              "type": "string",
              "maxLength": 255
            },
            "description": "A retry with the same key within 24 hours gets the first reply back (with `Idempotent-Replayed: true`) instead of running again."
          }
        ]
      }
    },
...
  },
...
  "components": {
    "securitySchemes": {
      "apiKey": {
        "type": "http",
        "scheme": "bearer",
        "description": "An API key: `tp_…`."
      }
    },
    "responses": {
...
      "Unauthorized": {
        "description": "`unauthorized`: no key, or it is unknown, revoked or expired.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "examples": {
              "error": {
                "value": {
                  "error": {
                    "code": "unauthorized",
                    "message": "…"
                  }
                }
              }
            }
          }
        }
      },
      "Forbidden": {
        "description": "`missing_scope`: the key lacks the scope this needs.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "examples": {
              "error": {
                "value": {
                  "error": {
                    "code": "missing_scope",
                    "message": "…"
                  }
                }
              }
            }
          }
        }
      },
      "NotFound": {
        "description": "`not_found`.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "examples": {
              "error": {
                "value": {
                  "error": {
                    "code": "not_found",
                    "message": "…"
                  }
                }
              }
            }
          }
        }
      },
      "Invalid": {
        "description": "`validation_failed`: `fields` says what is wrong with which input.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "examples": {
              "error": {
                "value": {
                  "error": {
                    "code": "validation_failed",
                    "message": "…"
                  }
                }
              }
            }
          }
        }
      },
      "RateLimited": {
        "description": "`rate_limited`: wait for Retry-After seconds.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "examples": {
              "error": {
                "value": {
                  "error": {
                    "code": "rate_limited",
                    "message": "…"
                  }
                }
              }
            }
          }
        },
        "headers": {
          "Retry-After": {
            "schema": {
              "type": "integer"
            }
          }
        }
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "object",
            "required": [
              "code",
              "message"
            ],
            "properties": {
              "code": {
                "type": "string"
              },
              "message": {
                "type": "string"
              },
              "fields": {
                "type": "object",
                "additionalProperties": {
                  "type": "string"
                }
              }
            }
          }
        }
      },
      "Traffic": {
        "type": "object",
        "required": [
          "rx_bytes",
          "tx_bytes",
          "total_bytes"
        ],
        "description": "Seen from the server: rx is what the device uploaded, tx what it downloaded.",
        "properties": {
          "rx_bytes": {
            "type": "integer"
          },
          "tx_bytes": {
            "type": "integer"
          },
          "total_bytes": {
            "type": "integer"
          }
        }
      },
      "DeviceStatus": {
        "type": "string",
        "enum": [
          "active",
          "disabled",
          "expired",
          "limit_reached"
        ]
      },
      "LimitPeriod": {
        "type": "string",
        "enum": [
          "monthly",
          "total"
        ],
        "description": "monthly counts from the first of each month in the panel's time zone; total counts until a usage reset."
      },
      "Device": {
        "type": "object",
        "required": [
          "id",
          "server_id",
          "name",
          "external_id",
          "metadata",
          "address",
          "public_key",
          "client_key",
          "enabled",
          "status",
          "online",
          "last_handshake",
          "data_limit",
          "limit_period",
          "period_usage",
          "usage_reset_at",
          "expires_at",
          "speed_limit",
          "month_usage",
          "created_at",
          "updated_at"
        ],
        "properties": {
          "id": {
            "type": "integer"
          },
          "server_id": {
            "type": "integer"
          },
          "name": {
            "type": "string"
          },
          "external_id": {
            "type": "string",
            "description": "Your own ID for the device's owner. Not unique."
          },
          "metadata": {
            "type": "object",
            "additionalProperties": true
          },
          "address": {
            "type": "string",
            "description": "The device's address inside the tunnel.",
            "examples": [
              "10.8.0.2"
            ]
          },
          "public_key": {
            "type": "string"
          },
          "client_key": {
            "type": "boolean",
            "description": "The client made the key pair; the panel never saw the private key."
          },
          "enabled": {
            "type": "boolean"
          },
          "status": {
            "$ref": "#/components/schemas/DeviceStatus"
          },
          "online": {
            "type": "boolean",
            "description": "As of the panel's last look, at most ten seconds old."
          },
          "last_handshake": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "data_limit": {
            "type": "integer",
            "minimum": 0,
            "description": "Bytes per limit period, both directions; 0 means no limit."
          },
          "limit_period": {
            "$ref": "#/components/schemas/LimitPeriod"
          },
          "period_usage": {
            "allOf": [
              {
                "$ref": "#/components/schemas/Traffic"
              }
            ],
            "description": "What counts toward data_limit: this month, or since the last usage reset."
          },
          "usage_reset_at": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time",
            "description": "The last usage reset, if any."
          },
          "expires_at": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time",
            "description": "Access ends at this instant."
          },
          "speed_limit": {
            "type": "integer",
            "minimum": 0,
            "description": "kbit/s (1000 bits a second), download and upload each; 0 means no limit."
          },
          "month_usage": {
            "allOf": [
              {
                "$ref": "#/components/schemas/Traffic"
              }
            ],
            "description": "The calendar month, whatever the limit period."
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          },
          "config": {
            "type": "string",
            "description": "Only in the reply to a create: the WireGuard config. Holds the private key unless public_key was sent."
          }
        }
      },
...
      "Group": {
        "type": "object",
        "required": [
          "external_id",
          "device_count",
          "month_usage",
          "devices"
        ],
        "properties": {
          "external_id": {
            "type": "string"
          },
          "device_count": {
            "type": "integer"
          },
          "month_usage": {
            "allOf": [
              {
                "$ref": "#/components/schemas/Traffic"
              }
            ],
            "description": "All the devices together, this calendar month."
          },
          "devices": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Device"
            }
          }
        }
      },
...
    }
  }
}